requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
mysqlclient==2.2.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
